It works 2.0

+++ src/com/company/classes/Worker.java
@@ -18,7 +18,7 @@ public class Worker implements Serializable {
 
     private static final long serialVersionUID = 66L;
 
-    private int id; //Поле не может быть null, Значение поля должно быть больше 0, Значение этого поля должно быть уникальным, Значение этого поля должно генерироваться автоматически
+    private long id; //Поле не может быть null, Значение поля должно быть больше 0, Значение этого поля должно быть уникальным, Значение этого поля должно генерироваться автоматически
     private String name; //Поле не может быть null, Строка не может быть пустой
     private Coordinates coordinates; //Поле не может быть null
     private java.time.ZonedDateTime creationDate; //Поле не может быть null, Значение этого поля должно генерироваться автоматически
@@ -29,7 +29,7 @@ public class Worker implements Serializable {
     private Person person; //Поле может быть null
     private User user;
 
-    public Worker(int ID, String name, double salary, Position position, Person person, Coordinates coordinates,
+    public Worker(long ID, String name, double salary, Position position, Person person, Coordinates coordinates,
                   ZonedDateTime startDate, ZonedDateTime endDate, User user) throws InvalidDataException{
         setName(name);
         setSalary(salary);
@@ -183,7 +183,7 @@ public class Worker implements Serializable {
 
     //getters and setters
 
-    public int getId() {
+    public long getId() {
         return id;
     }
